+++ src/Python/Sample Data/qrcodes/simple_openMap.png
